Ensure strncmp compare include null-terminate char
If we use n-paramater of strncmp as strlen, it will check only
Prefix of string.

For example,
char a[] = "asdf";
char b[] = "as";

then, strncmp(a, b, strlen(b)) will return 0, even if a and b is not equal.

To avoid this kind of problem, let we ensurely check end of comparitor is
delim or not.

Change-Id: I681cded5a85252f4b882ec480d1c4bc961d2c9a2
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>