Make ShaderPrecompiler move shader list + Store shader prefix list as string instead string_view
Until now, Precompiler hold only string_view.
It mean, the lifecycle of shader code is not depend on that object.

To make sure that we could control the lifecycle of source code string,
let we keep it as std::string, instead of std::string_view.

And also, make we use std::move instead of copy them.
It will make less performance overhead.

Change-Id: I2b5949b5a408865b5879448630c65a6d3682ecee
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>