Ensure we don't touch shutdowned EglGraphicsController
Let we make IsShuttingDown() API as static,
so we can ask it without raw pointer of graphics controller.

Most of API in GLES::Resource hold the eglController as l-value.
So if someone try to get graphics controller after shutting down,
it might make some crash.

To avoid these kind of issue, let we ensure to call DiscardResource
only if graphics controller is not in shutting down.

TODO : native image should be called DiscardResource() even if it is shutting
downed. How can we ensure it?

Change-Id: I87ec5d96d8c12c054f00db346564f7bae8d5d3bb
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>