Free sync object immediately if current context is same as created context
Until now, we usually destroy sync object at surfaceless context
at the end of rendering.

But it might have some context switching overhead occured for some driver.

To avoid it, let we try to free sync object if read context is
same as write context, and it is already synced.

Change-Id: Ibec3cb40379ad341ff757362ee6cb529806153a3
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>