Swapchain improvement
In some cases, usually during startup, we acquire a swapchain image but never consume it because there is no render pass started. This leaves acquire image semaphores in the signaled state and breaks synchronization when same image is about to be acquire frames later (depends on driver, it may be ignored or it may lead to deadlock).

Since we assume that only one render pass starts for each surface we acquire the image only when such render pass starts (so we have something to render). This also removes "default" AcquireNextFramebuffer() function in the controller which assumes we have a single window only.

The infinite acquire timeout has been replaced with arbitrary 1000000000 value which makes sure we won't get into the deadlock state (vkAcquireNextImageKHR() will return timeout error) but should be enough to acquire next image.

The assert that used to crash the demos happens only on return of error code (timeout and not_ready aren't errors).

Change-Id: Idd872c8fcde80ecb1f5c71fd587564e210ea3a9d