N buffering vulkan resource discard queues
As we now have multiple command buffers per frame, it is important
to only discard resources for old frames; we run the GC every time
Submit is called.

So, we need a discard q per buffer index; which may be 0,1 or 0,1,2
depending on platform.

Added a vector of queues that gets automatically allocated as needed.

Change-Id: I724e5d3ae3c5b15f99ab2821ef1c75944b9e24cb
Signed-off-by: David Steele <david.steele@samsung.com>