Re-factoring vk discard queues.
Changed Resource, so now we have ResourceWithDeleter
and ResourceWithoutDeleter, but ResourceBase can be used
as common type with calls to GetAllocationCallbacks and
InvokeDeleter.

(Note, only Program/Pipeline objects currently inherit
from ResourceWithoutDeleter, as these are proxy objects
into a cache)

Means that the DiscardQueue is simplified so it only
deals with ResourceBase objects. (Don't have to do
specialization magic depending on resource type).

Change-Id: If8538fa9bbdae9fb6f5b3aeb183c25b9523c673b