Fix rarely crash if NativeImage discarded after draw occured
It is possible s.t. DiscardNativeImage called before EndRenderPass called.

Until now, we remove the container directly. If this case,
the count of native texture dependency might be zero, so EndRenderPass logic
might be skipped!

To avoid this case, let we don't touch the contianer itself, and just mark
it as synced.

Change-Id: I3171493c0d72ccd58b7bdcacac47876ea32ed1d9
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>