Ensure empty render instruction doesn't swap buffers
If we aren't going to draw anything to the surface, then
don't acquire next image or render anything.

SwapBuffers is handled through Swapchain::Present, which maintains
it's own framecount; The current buffer index is retrieved from the
main window's swapchain.

Updated VulkanGraphicsController::PresentRenderTarget to also call
WindowSurface::PostRender; this has a load of handling for partial
rendering that we may need in the future.

Will keep rendering framebuffers, regardless of whether they are drawn
or not.

Change-Id: I8e0b2555647e2ce74d425c9632246f19ea295470