[Tizen] Core::PreRender returns true if something is renderable to scene
This is a combination of 5 commits.

Ensure empty render instruction doesn't swap buffers

If we aren't going to draw anything to the surface, then
don't acquire next image or render anything.

SwapBuffers is handled through Swapchain::Present, which maintains
it's own framecount; The current buffer index is retrieved from the
main window's swapchain.

Updated VulkanGraphicsController::PresentRenderTarget to also call
WindowSurface::PostRender; this has a load of handling for partial
rendering that we may need in the future.

Will keep rendering framebuffers, regardless of whether they are drawn
or not.

Change-Id: I8e0b2555647e2ce74d425c9632246f19ea295470

Debug extra

Change-Id: I05c8b768acfdebfce8fa661ff69e6e13d6412a53

Reworking pre-render/post-render

Changed main render loop to check if there is a full-swap
without any rendering (usually caused by having no render-task
yet, but resizing or setting the background color of the window).

If so, we need to ensure that an empty scene can be correctly
rendered. Added a ClearScene to Core API.

This is more pertinent to Vulkan renderer, which has to ensure
that the window's render surface is properly acquired _before_
drawing to it, and presenting that surface needs the right
sync primitives.

Modified VulkanGraphicsController::PresentRenderTarget() to also call
Surface->PostRender(), in the same way that EglGraphicsController does.
(Prelude to other re-work for PartialRendering).

Change-Id: Ic3fb46faa17b3ec67053cfc4fbd48f6a31cc414c

Fixing swapchain buffer count and cmd buffer allocs

Number of swapchain buffers should be based on minImageCount
from surfaceCapabilities, rather than the current number of
images that getSwapchainImagesKHR returns.

Currently, it is possible for the swapchain to get invalidated,
and the new swapchain can be generated with more images than
min. We only need to sync the min number; so we double or triple
buffer based on this count.

(number of CommandBuffers / SwapchainBuffers is based on this min count,
not total image count)

Changed this to only reallocate command buffers in Reset(), rather than
breaking any outstanding sequence.

Change-Id: I5348dcc1762426e5370a35cd7f48d4612a3b5114

Ensure that RenderScene actually renders something.

RenderScene should only get called if there is something to
present and nothing prevents a renderer being drawn.

The only early-out that's allowed now is null scene handle.

ClearScene will only get called if RenderScene did not present
when we expected it to have actually drawn something.

If there are no renderers, but we set the background color of
the window; currently, FullSwapNextFrame is set. If this is not
desired behaviour, it should be changed in
SceneHolder::SetBackgroundColor.

Change-Id: I549d2fd5a25f6a182d89c8c2cc5abfaa07f985c6
Signed-off-by: David Steele <david.steele@samsung.com>