Reworking pre-render/post-render
Changed main render loop to check if there is a full-swap
without any rendering (usually caused by having no render-task
yet, but resizing or setting the background color of the window).

If so, we need to ensure that an empty scene can be correctly
rendered. Added a ClearScene to Core API.

This is more pertinent to Vulkan renderer, which has to ensure
that the window's render surface is properly acquired _before_
drawing to it, and presenting that surface needs the right
sync primitives.

Modified VulkanGraphicsController::PresentRenderTarget() to also call
Surface->PostRender(), in the same way that EglGraphicsController does.
(Prelude to other re-work for PartialRendering).

Change-Id: Ic3fb46faa17b3ec67053cfc4fbd48f6a31cc414c