Fixing swapchain buffer count and cmd buffer allocs
Number of swapchain buffers should be based on minImageCount
from surfaceCapabilities, rather than the current number of
images that getSwapchainImagesKHR returns.

Currently, it is possible for the swapchain to get invalidated,
and the new swapchain can be generated with more images than
min. We only need to sync the min number; so we double or triple
buffer based on this count.

(number of CommandBuffers / SwapchainBuffers is based on this min count,
not total image count)

Changed this to only reallocate command buffers in Reset(), rather than
breaking any outstanding sequence.

Change-Id: I5348dcc1762426e5370a35cd7f48d4612a3b5114