Ensure that RenderScene actually renders something.
RenderScene should only get called if there is something to
present and nothing prevents a renderer being drawn.

The only early-out that's allowed now is null scene handle.

ClearScene will only get called if RenderScene did not present
when we expected it to have actually drawn something.

If there are no renderers, but we set the background color of
the window; currently, FullSwapNextFrame is set. If this is not
desired behaviour, it should be changed in
SceneHolder::SetBackgroundColor.

Change-Id: I549d2fd5a25f6a182d89c8c2cc5abfaa07f985c6
Signed-off-by: David Steele <david.steele@samsung.com>