ClearScene if previous frame presented + Present if dirty rect not empty
Their was some error if rendererAdded, but render instruction is empty.

If then, we don't swap the presented buffer. So latest rendered scene
don't cleared.

To fix this issue, let we acquire next frame if previous scene had some render
instructions to scene, but not now.

Also (for gles case) if we call eglSetDamageRegionKHR, we should call eglSwapBuffer.
If not, 0x3002 error occured.

Change-Id: Icf55be6fe92ab98def0e631b0205939c16a8f4e6
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>