Refactor: font face cache
The core change in this patch is to create 1 freetype face per font file.
Previously, we created as many faces as font-face-cache-items.

This was fine for static (seldom changing) apps, but when using a wide variety of point sizes,
it would create a face every time, which used too much memory.

Now we reuse 1 face.
Instead of faces, freetype sizes are cached and activated on demand,
which uses significantly less memory than the previous face cache.

In static apps, the memory usage difference may not be significant,
but in extreme cases where point sizes exceed thousands, it is incomparably efficient.

And now, we can use pre-loaded memory fonts in async text as well via the font file manager!

TODO: Optimize font-face-cache-item.

Change-Id: I38f44e641e00ae7bfcf16891f92ba14d74c9bff9
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>