(VectorImage) Ensure to load completed at IsLoaded()
Until now, we return IsLoaded() value by some memory pointer.
But IsLoaded could be called from any thread, so sometimes
it return true even if it is invalid resouce.

To avoid this case, let we use atomic bool, so we can know
it load completed at the end of Load() API.

Change-Id: I8625c29ff8eccf7f871defa59e1e8fab86001389
Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>