Call glInvalidateDepthStencilRenderBuffer at EndRenderPass
Until now, we call InvalidateFramebuffer for 0 framebuffer,
when we call PresentRenderTarget().

The behavior of this API changed at 10.0. Also this invalidate should
be called "Before" the custom framebuffer changed to other framebuffer.

reference : https://community.arm.com/developer/tools-software/graphics/b/blog/posts/mali-performance-2-how-to-correctly-handle-framebuffers

So we should call it at the end of renderpass, if the framebuffer don't
have render target textures.

Change-Id: Ic126d916b4f9f49a3710c575360d1b17ce2fc736
Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>