Ensure to blend and cullface cache value are really call gl function
If pipeline comes like below flow, cached gl blend status not matched with real gl blend status.

Blend : ON, BlendFuncAndOp : A  --> Cache as ON and A
Blend : OFF, BlendFuncAndOp : B --> Cache as OFF and A (since Off blend early-out so cache not chaneed)
Blend : ON, BlendFuncAndOp : B --> Cache as ON and A (since BlendFuncAndOp is same as previous)

To avoid this kind of issue, we mark whether this pipeline is the first times or not.

And also, let we don't check previous color state anymore.
So the cached blend status must be same as real gl statue now.

Change-Id: Ie7cc3bf72ad5a99f0b8aaca3c310eb49ddf92464
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>