Second window doesn't overwrite Vk device
Surface initialization is responsible for creating the
vulkan physical and logical devices.

We only need 1 physical device (and probably also only 1
logical device for now), so prevent a 2nd window overriding
the devices.

In the future, we may want to consider separating the logical
device and queues per DeviceWindow (Renamed SurfaceSwapchainPair).

Change-Id: Ife52bb5ef987e9eea722a16405eaa324f3e4cd0e
Signed-off-by: David Steele <david.steele@samsung.com>