Fixed the unsafe mode rendering issue in DRGL View
When using the unsafe mode in direct rendering mode, it initializes the mActiveTextureUnit value in the cache.
however, since the initialized value is 0, there’s an issue where glActiveTexture isn’t called when the index is 0.
To fix this, we’ve modified it so that it's initilized value is max.

Change-Id: I3145446b621c0c0de0513db3ff5613c5ee437f47