Share ownership of Accessibility::Bridge ownership to adaptor and application
If static Dali::Application::PreInitializedApplication terminated,
the order of static variables are not determined.

Let we ensure to keep ownership of Bridge, so we can ensure to call Terminate()
at Adaptor destructor.

Also, we should keep the reference of bridge in application too,
since we can try to destructor application without adaptor creation.

Change-Id: Ibbb31925d6dff5581c6d578c6aefdb27fa8bc01d
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>