Adjust the creation sequence of HarfBuzzProxyFont
There is an issue where all inputs are identical,
but the hb_shape results are slightly different.

Based on multiple tests,
there appears to be a relation between the harfbuzz font and the size of ftFace.

And according to the reference,
the hb_shape results should be identical in the two sequences below.

1. harfbuzz font creation -> ftFace size change -> hb_ft_font_changed -> shape
2. ftFace size chanage -> harfbuzz font creation (No need to call hb_ft_font_changed) -> shape

But case 1 sometimes doesn't seem to work well.

This fix eliminated the issue,
but I'm still not sure if this is a harfbuzz or freetype issue.

+ Explicitly call hb_ft_font_changed and hb_ft_font_set_load_flags.

Change-Id: I55ba5c86f1f0b1eb10cc3c5e3a4e83f992d077c9
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>