(Vulkan) Prevent swapchain lockup during Submit
If there are multiple render tasks without fbos (e.g. multiple
views onto the surface), then we were creating several submissions
waiting for the same sema, which would never happen.

Changed the submission so that multiple command buffers can be
submitted for the surface in a single Queue::Submit, and set up
the semaphores appropriately.

Change-Id: I67632eec4ad7b22820775465d69ed8f2b885a7dd
Signed-off-by: David Steele <david.steele@samsung.com>