(Vulkan) Adding texture array for fbo batching
During FBO setup in any given frame, we check all
color attachment vulkan image creation, and defer it
until just before rendering. Then, we batch up all
images of the same size into as few image arrays as
possible.

The image array is owned by a Texture as usual,
but this is wrapped with a TextureArray object, which
maintains a refcount of all batched attachments.
If this refcount drops to zero, then the image array
will be removed.

Due to odd handling in Control, fbos may be created and
destroyed in the same frame. We ensure that no
redundant image/image view pairs are batched.

Change-Id: Iedf1bd03f31ce1a056f507990dc3999e73d71d45