[Tizen] Adding Dimming Functionality to Window Behind Blur
This patch introduces a new dimming feature for the "Window Behind Blur" functionality,
enhancing the visual experience by allowing users to adjust the opacity or brightness of the blurred background window.
The dimming feature provides greater control over the appearance of the blurred window,
making it more adaptable to various user preferences and use cases.

Change-Id: If924ceae9c70b53019ceab81b726c428d205e11d