(ConfigurationManager) Seperate cache file per each graphics backend
Very rare case but, If we delete and re-write the gpu configure for
both Vulkan / GLES backend from multiple application, the cached result
might be breakdown.

Let we ensure to do not use, or write to cache file if backend is not matched with
cache file. And let we only touch the matched graphics backend file as
Initialized time.

Also, minimize the file stream opend time, so could reduce race condition of file.

Change-Id: Ieb9d1340c53d46d1ca55d62cdaec681fe3263237
Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>